requirements.txt
fastapi
uvicorn
pydantic
python-dateutil
starlette
flake8
isort
httpx
pytest
pytest-cov